third_party/blink/web_tests/platform/linux-chrome/external/wpt/secure-payment-confirmation/enrollment-in-iframe.sub.https-expected.txt

This is a testharness.js-based test.
[FAIL] SPC enrollment in cross-origin iframe
  assert_equals: expected null but got object "NotSupportedError: The user agent does not support public key credentials."
Harness: the test ran to completion.
